# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.DefaultResponse:
    properties:
      message:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Company:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.SwagVacancy:
    properties:
      company_id:
        type: integer
      description:
        type: string
      location:
        type: string
      salary:
        type: number
      title:
        type: string
      vacancy_category_id:
        type: integer
    type: object
...
  models.User:
    properties:
      birth_date:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      username:
        type: string
    type: object
  models.Vacancy:
    properties:
      company:
        $ref: '#/definitions/models.Company'
      company_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      salary:
        type: number
      title:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      vacancy_category_id:
        type: integer
      vacancyCategory:
        $ref: '#/definitions/models.VacancyCategory'
      vacancyViews:
        items:
          $ref: '#/definitions/models.VacancyView'
        type: array
    type: object
  models.VacancyCategory:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.VacancyReport:
    properties:
      applications_count:
        type: integer
      vacancy_id:
        type: integer
      vacancy_title:
        type: string
      views_count:
        type: integer
    type: object
  models.VacancyView:
    properties:
      count:
        type: integer
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      vacancy:
        $ref: '#/definitions/models.Vacancy'
      vacancy_id:
        type: integer
    type: object
host: localhost:8181
info:
  contact: {}
  description: This is a Tajik Career Hub API documentation.
  title: Tajik Career Hub API ✨
  version: "1.0"
paths:
...
  /activities/vacancy/{id}:
    get:
      consumes:
      - application/json
      description: Get a report of how many people viewed or applied to a specific
        vacancy
      operationId: get-vacancy-report-by-id
      parameters:
      - description: Vacancy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VacancyReport'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get report for a specific vacancy
      tags:
      - Reports
...
  /vacancies:
    get:
      consumes:
      - application/json
      description: Get a list of all vacancies with optional filters such as search,
        salary range, location, category, and sort order.
      operationId: get-all-vacancies
      parameters:
      - description: Search keyword for filtering vacancies
        in: query
        name: search
        type: string
      - description: Minimum salary for filtering vacancies
        in: query
        name: minSalary
        type: integer
      - description: Maximum salary for filtering vacancies
        in: query
        name: maxSalary
        type: integer
      - description: Location for filtering vacancies
        in: query
        name: location
        type: string
      - description: Category for filtering vacancies
        in: query
        name: category
        type: string
      - description: Sorting order for vacancies
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of vacancies
          schema:
            items:
              $ref: '#/definitions/models.Vacancy'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all vacancies with filters
      tags:
      - Vacancies
    post:
      consumes:
      - application/json
      description: Add a new vacancy with the provided details.
      operationId: add-vacancy
      parameters:
      - description: User ID to check if the user is blocked
        in: query
        name: userID
        required: true
        type: integer
      - description: Vacancy object to be added
        in: body
        name: vacancy
        required: true
        schema:
          $ref: '#/definitions/models.SwagVacancy'
      produces:
      - application/json
      responses:
        "201":
          description: Vacancy created successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: ErrPermissionDenied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new vacancy
      tags:
      - Vacancies
  /vacancies/{vacancyID}:
    delete:
      consumes:
      - application/json
      description: Soft delete a specific vacancy by its ID.
      operationId: delete-vacancy
      parameters:
      - description: User ID to check if the user is blocked
        in: query
        name: userID
        required: true
        type: integer
      - description: ID of the vacancy to delete
        in: path
        name: vacancyID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Vacancy deleted successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Vacancy Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a vacancy
      tags:
      - Vacancies
    get:
      consumes:
      - application/json
      description: Get details of a single vacancy by its ID.
      operationId: get-vacancy-by-id
      parameters:
      - description: ID of the vacancy to retrieve
        in: path
        name: vacancyID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved vacancy
          schema:
            $ref: '#/definitions/models.SwagVacancy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Vacancy Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a specific vacancy by ID
      tags:
      - Vacancies
    put:
      consumes:
      - application/json
      description: Update an existing vacancy by its ID.
      operationId: update-vacancy
      parameters:
      - description: User ID to check if the user is blocked
        in: query
        name: userID
        required: true
        type: integer
      - description: ID of the vacancy to update
        in: path
        name: vacancyID
        required: true
        type: integer
      - description: Updated vacancy object
        in: body
        name: vacancy
        required: true
        schema:
          $ref: '#/definitions/models.SwagVacancy'
      produces:
      - application/json
      responses:
        "200":
          description: Vacancy updated successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Vacancy Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing vacancy
      tags:
      - Vacancies
  /vacancies/block/{id}:
    patch:
      consumes:
      - application/json
      description: Blocks a vacancy by its ID. Requires the user to be authenticated.
      parameters:
      - description: Vacancy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Vacancy blocked successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid vacancy ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Block a vacancy
      tags:
      - Admin
  /vacancies/unblock/{id}:
    patch:
      consumes:
      - application/json
      description: Unblocks a vacancy by its ID. Requires the user to be authenticated.
      parameters:
      - description: Vacancy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Vacancy unblocked successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid vacancy ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unblock a vacancy
      tags:
      - Admin
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"